requirements.txt

torchfile==0.1.0
torchsummary==1.5.1
numpy==1.21.2
nvidia_ml_py3==7.352.0
opencv_python==4.5.5.62
Pillow==9.1.1
pynvml==11.4.1
scipy==1.8.0
torchviz==0.0.2
torchwisdom==0.0.1
tqdm==4.62.3
datasets==2.0.0
huggingface-hub
timm==0.4.12
transformers==4.30.0
peft==0.11.1
pycocotools
ftfy
matplotlib==3.7.1
contourpy==1.1.1